fixed serviceworker with workbox
