+++ src/scripts/site.js
@@ -225,7 +225,11 @@ function registerServiceWorker() {
     /* eslint-disable compat/compat */
     if (navigator.serviceWorker && window.appMode !== 'cordova' && window.appMode !== 'android') {
         try {
-            navigator.serviceWorker.register('serviceworker.js');
+            navigator.serviceWorker.register('/serviceworker.js').then(() =>
+                console.log('serviceWorker registered')
+            ).catch(error =>
+                console.log('error registering serviceWorker: ' + error)
+            );
         } catch (err) {
             console.error('error registering serviceWorker: ' + err);
         }
